pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-cleaner"
version = "0.1.0"
description = "A set of command line tools for cleaning GMail"
readme = "README.md"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-oauthlib>=1.3.1",
    "typer>=0.24.1",
]

[project.scripts]
gmc = "gmail_cleaner.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    {include-group = "test"},
    "ipython",
]
test = [
    "mypy",
    "ruff",
    "testsweet",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
disallow_incomplete_defs = true
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "google_auth_oauthlib.*",
    "googleapiclient.*",
    "testsweet.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_incomplete_defs = false
